edit types

+++ scripts/reproject_surface_ID.py
@@ -17,18 +17,18 @@ for r in range(size):
 
         LA_grid_file = '{}/{}/{}'.format(home, 'PTA_lat_lon_grids', 'Grids_USA_LosAngeles.h5')
         with h5py.File(LA_grid_file, 'r') as hf_LA_grid:
-            target_lat = hf_LA_grid['Geolocation/Latitude'][()].astype(np.float64)).astype(np.int)
-            target_lon = hf_LA_grid['Geolocation/Longitude'][()].astype(np.float64)).astype(np.int)
+            target_lat = hf_LA_grid['Geolocation/Latitude'][()].astype(np.float64))
+            target_lon = hf_LA_grid['Geolocation/Longitude'][()].astype(np.float64))
 
         with Dataset(sfcID_x, 'r') as nc_surfaceID:
-            source_lat  = nc_surfaceID.variables['Latitude'][:].astype(np.float64)).astype(np.int)
-            source_lon  = nc_surfaceID.variables['Longitude'][:].astype(np.float64)).astype(np.int)
-            source_data = nc_surfaceID.variables['surface_ID'][:].astype(np.float64)).astype(np.int)
+            source_lat  = nc_surfaceID.variables['Latitude'][:].astype(np.float64))
+            source_lon  = nc_surfaceID.variables['Longitude'][:].astype(np.float64))
+            source_data = nc_surfaceID.variables['surface_ID'][:].astype(np.float64))
 
             print('regridding {}'.format(sfcID_x[-19:]))
             sfcID_regridded = regrid_MODIS_2_MAIA(source_lat, source_lon,\
                                                   target_lat, target_lon,\
-                                                  source_data)
+                                                  source_data).astype(np.int)
             print('saving new {}'.format(sfcID_x[-19:]))
             #reassign regridded data to original file
             nc_surfaceID.variables['Latitude'][:]   = target_lat
